+++ template_cpp/src/include/delivery_set.hpp
@@ -9,7 +9,7 @@
 class DeliverySet
 {
 private:
-    std::map<uint32_t, std::set<uint64_t>> delivered_messages;
+    std::map<uint32_t, std::set<uint8_t>> delivered_messages;
     HostLookup host_lookup;
 
 public:
@@ -17,10 +17,10 @@ public:
 
     bool is_delivered(TransportMessage message)
     {
-        uint64_t host_id = this->host_lookup.get_host_id_by_ip(message.address);
+        uint8_t host_id = this->host_lookup.get_host_id_by_ip(message.address);
         auto iterator = this->delivered_messages.find(message.get_id());
         auto set =
-            iterator == this->delivered_messages.cend() ? std::set<uint64_t>() : iterator->second;
+            iterator == this->delivered_messages.cend() ? std::set<uint8_t>() : iterator->second;
         bool found = !(set.find(host_id) == set.cend());
 
         return found;
@@ -28,10 +28,10 @@ public:
 
     void deliver(TransportMessage message)
     {
-        uint64_t host_id = this->host_lookup.get_host_id_by_ip(message.address);
+        uint8_t host_id = this->host_lookup.get_host_id_by_ip(message.address);
         if (this->delivered_messages.find(message.get_id()) == this->delivered_messages.cend())
         {
-            std::set<uint64_t> set;
+            std::set<uint8_t> set;
             set.insert(host_id);
             this->delivered_messages.emplace(message.get_id(), set);
             return;
